Use public_name for vosk-eio library
* What

This commit updates the `vosk-eio` library definition in the `dune`
file. Specifically, it changes the `package` field to `public_name`.

* Why

The `public_name` field is used to define how a library is exposed and
made available for other projects to depend on or for installation.
Using `public_name` ensures that the `vosk-eio` library can be properly
installed and referenced by other packages or users. This is a standard
practice for libraries intended for public use within the OCaml
ecosystem.

* How

The change was made by replacing `(package vosk-eio)` with `(public_name
vosk-eio)` in the `src/eio/dune` file. This is a direct modification
within the library stanza.